refactor(tests): rename feature tests to end-to-end tests
