+++ src/test/e2e-emulated/contracts/dns.tact
