+++ src/test/e2e-emulated/contracts/implicit-init.tact
