+++ src/test/e2e-emulated/contracts/integer-literals.tact
