+++ src/test/e2e-emulated/contracts/intrinsics.tact
