+++ src/test/e2e-emulated/contracts/map-traverse.tact
