+++ src/test/e2e-emulated/contracts/masterchain.tact
