+++ src/test/e2e-emulated/contracts/serialization-2.tact
