+++ src/test/e2e-emulated/contracts/serialization-3.tact
