+++ src/test/e2e-emulated/contracts/serialization.tact
