+++ src/test/e2e-emulated/contracts/stdlib.tact
