+++ src/test/e2e-emulated/contracts/strings.tact
