+++ src/test/e2e-emulated/contracts/ternary.tact
