+++ src/test/e2e-emulated/feature-address.spec.ts
@@ -1,8 +1,8 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { AddressTester } from "./contracts/output/address_AddressTester";
-import { consoleLogger } from "../logger";
+import { consoleLogger } from "../../logger";
 
 describe("feature-address", () => {
     beforeAll(() => {
