+++ src/test/e2e-emulated/feature-bounced-routing.spec.ts
@@ -1,6 +1,6 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { SampleContract2 } from "./contracts/output/bounced-routing_SampleContract2";
 import { SampleContract } from "./contracts/output/bounced-routing_SampleContract";
