+++ src/test/e2e-emulated/feature-constants.spec.ts
@@ -1,6 +1,6 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { ConstantTester } from "./contracts/output/constants_ConstantTester";
 
 describe("feature-constants", () => {
