+++ src/test/e2e-emulated/feature-dns.spec.ts
@@ -1,6 +1,6 @@
 import { OpenedContract, beginCell, toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { DNSTester } from "./contracts/output/dns_DNSTester";
 
 function convertToInternal(src: string) {
