+++ src/test/e2e-emulated/feature-implicit-init.spec.ts
@@ -1,6 +1,6 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { MyContract } from "./contracts/output/implicit-init_MyContract";
 
 describe("feature-implicit-init", () => {
