+++ src/test/e2e-emulated/feature-intrinsics.spec.ts
@@ -1,6 +1,6 @@
 import { Address, beginCell, Cell, toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { IntrinsicsTester } from "./contracts/output/intrinsics_IntrinsicsTester";
 import { sha256_sync } from "@ton/crypto";
