+++ src/test/e2e-emulated/feature-ordering.spec.ts
@@ -1,6 +1,6 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { A } from "./contracts/output/ordering_A";
 import { B } from "./contracts/output/ordering_B";
