+++ src/test/e2e-emulated/feature-strings.spec.ts
@@ -1,6 +1,6 @@
 import { beginCell, toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { StringsTester } from "./contracts/output/strings_StringsTester";
 
 describe("feature-strings", () => {
