+++ src/test/e2e-emulated/feature-try-catch.spec.ts
@@ -1,6 +1,6 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
-import { __DANGER_resetNodeId } from "../grammar/ast";
+import { __DANGER_resetNodeId } from "../../grammar/ast";
 import { TryCatchTester } from "./contracts/output/try-catch_TryCatchTester";
 
 describe("feature-ternary", () => {
